reorganize fitting module.
