+++ prospect/fitting/__init__.py
@@ -1,4 +1,7 @@
-from .fitterutils import *
+from .ensemble import *
+from .minimizer import *
+from .nested import *
 
 __all__ = ["run_emcee_sampler", "reinitialize_ball", "sampler_ball",
+           "run_nested_sampler",
            "pminimize", "minimizer_ball", "reinitialize"]
